Update UnitCost and add Quantity to InventoryBatchActivity
This commit updates the `UnitCost` for specific entries in the `finance.InventoryBatch` table based on `PurchaseDate` and `ProductTypeId`. A new entry is added to set the `UnitCost` for a product with `ProductTypeId = 6` on '14 Jun 2025'. Additionally, the `INSERT` statement in the `finance.InventoryBatchActivity` table is modified to include a new `Quantity` column, enabling the recording of the quantity associated with each activity.