Rename QuantityPurchased to InitialQuantity
Updated SQL files to reflect the renaming of the
`QuantityPurchased` column to `InitialQuantity` in the
`finance.InventoryBatch` table. Adjusted queries in
`Populate InventoryBatch Data.sql` for consistency.
Added a new operation in the refactor log to document
this change. Included `Populate Transactions.sql` in
the project and added multiple new transaction entries
executing the `Transaction_CreateExpenseIncome` stored
procedure.