Enhance StockSale UI and add TotalQuantity property
Added a new UI section in `StockSale.razor` to display the
"Total Quantity" when multiple items exist in the
`EditStockSale.DetailList`. Introduced a computed property
`TotalQuantity` in `StockSaleBase` to calculate the sum of
quantities in the list.

Improved the `StockSaleList.razor` UI by styling "No" statuses
for `SaleConfirmed` and `PaymentReceived` with a red background
and rounded corners for better visual distinction.